Refactor matching service
