+++ matching-service/index.js
@@ -5,7 +5,7 @@ import { createServer } from 'http';
 import { Server } from 'socket.io';
 import axios from 'axios';
 import {createMatch, deleteMatch } from './controller/match-controller.js';
-import { ormCreateMatch } from './service/match-orm.js';
+import { ormCreateMatch } from './model/match-orm.js';
 
 const app = express();
 const server = createServer(app);
